Fixed mp bugs
- Added a timeout for all the queues. This can be changed in
SpInterpMain.

- Changed some names with Krige in them to Spinterp.

- Drift arrays, station drifts and vg_ser are passed to SpInterpSteps if
the type of interpolation is compatible.

- Next step is to replace krige with interp in all variable names.

+++ interp/drift.py
@@ -77,7 +77,7 @@ class KrigingDrift:
             drift_cols = drift_ds.RasterXSize
 
             drift_band = drift_ds.GetRasterBand(1)
-            drift_arr = drift_band.ReadAsArray()
+            drift_arr = drift_band.ReadAsArray().astype(float)
 
             self._drft_oarrs.append(drift_arr)
